requirements.txt
aiohttp~=3.8.4
asyncpg
SQLAlchemy
python-dotenv